rename test and add update test
